processor: Refactor PluginsContainer into a top level class
