+++ biz.aQute.bndlib/src/aQute/bnd/osgi/Processor.java
@@ -1,7 +1,5 @@
 package aQute.bnd.osgi;
 
-import static java.lang.invoke.MethodHandles.publicLookup;
-import static java.lang.invoke.MethodType.methodType;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toList;
@@ -13,7 +11,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.PrintStream;
-import java.lang.invoke.MethodType;
 import java.lang.reflect.Array;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
@@ -24,20 +21,17 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.attribute.FileTime;
-import java.util.AbstractSet;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Properties;
 import java.util.Random;
 import java.util.Set;
@@ -45,7 +39,6 @@ import java.util.Spliterator;
 import java.util.TreeSet;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.concurrent.CopyOnWriteArraySet;
 import java.util.concurrent.Executor;
 import java.util.concurrent.Executors;
 import java.util.concurrent.RejectedExecutionHandler;
@@ -80,15 +73,12 @@ import aQute.bnd.memoize.CloseableMemoize;
 import aQute.bnd.memoize.Memoize;
 import aQute.bnd.service.Plugin;
 import aQute.bnd.service.Registry;
-import aQute.bnd.service.RegistryDonePlugin;
 import aQute.bnd.service.RegistryPlugin;
-import aQute.bnd.service.url.URLConnectionHandler;
 import aQute.bnd.stream.MapStream;
 import aQute.bnd.unmodifiable.Lists;
 import aQute.bnd.version.Version;
 import aQute.bnd.version.VersionRange;
 import aQute.lib.collections.Iterables;
-import aQute.lib.hex.Hex;
 import aQute.lib.io.IO;
 import aQute.lib.io.IOConstants;
 import aQute.lib.strings.Strings;
@@ -102,7 +92,7 @@ import aQute.libg.reporter.ReporterAdapter;
 import aQute.service.reporter.Reporter;
 
 public class Processor extends Domain implements Reporter, Registry, Constants, Closeable {
-	static final Logger		logger	= LoggerFactory.getLogger(Processor.class);
+	private static final Logger	logger	= LoggerFactory.getLogger(Processor.class);
 	public static Reporter		log;
 
 	static {
@@ -564,7 +554,7 @@ public class Processor extends Domain implements Reporter, Registry, Constants, 
 	private Memoize<PluginsContainer> newPluginsContainer() {
 		Memoize<PluginsContainer> supplier = Memoize.supplier(() -> {
 			PluginsContainer pc = new PluginsContainer();
-			pc.init();
+			pc.init(this);
 			return pc;
 		});
 		// do postInit outside of above memoizer to allow reentrant
@@ -573,383 +563,11 @@ public class Processor extends Domain implements Reporter, Registry, Constants, 
 		return Memoize.predicateSupplier(supplier, pc -> {
 			// only first caller gets to do postInit
 			if (postInit.getAndSet(false)) {
-				pc.postInit();
+				pc.postInit(this);
 			}
 			return true; // always memoize the PluginsContainer
 		});
 	}
-
-	/**
-	 * Member class which contains the current plugin set for the Processor.
-	 */
-	public class PluginsContainer extends AbstractSet<Object> implements Set<Object>, Registry {
-		private final Set<Object>			plugins				= new CopyOnWriteArraySet<>();
-		// The following are only mutated during init(), so they don't need to
-		// be concurrent-safe
-		private final Set<String>			missingCommand		= new HashSet<>();
-		private final Set<AutoCloseable>	closeablePlugins	= new HashSet<>();
-
-		protected PluginsContainer() {}
-
-		/**
-		 * Init actions occur inside of the first-level memoizer.
-		 */
-		protected void init() {
-			String spe = getProperty(PLUGIN);
-			if (NONE.equals(spe)) {
-				return;
-			}
-
-			// The owner of the plugin is always in there.
-			add(Processor.this);
-			setTypeSpecificPlugins(this);
-
-			if (getParent() != null) {
-				addAll(getParent().getPlugins());
-			}
-
-			//
-			// Look only local
-			//
-
-			spe = mergeLocalProperties(PLUGIN);
-			String pluginPath = mergeProperties(PLUGINPATH);
-			loadPlugins(spe, pluginPath);
-		}
-
-		/**
-		 * Post init actions must occur outside of the first level memoizer.
-		 * This means these actions can reentrantly see the current state of the
-		 * PluginsContainer, through the Processor, which may be partially
-		 * complete if addExtensions adds more plugins.
-		 */
-		protected void postInit() {
-			addExtensions(this);
-
-			for (RegistryDonePlugin rdp : getPlugins(RegistryDonePlugin.class)) {
-				try {
-					rdp.done();
-				} catch (Exception e) {
-					exception(e, "Calling done on %s, gives an exception", rdp);
-				}
-			}
-		}
-
-		protected Set<Object> plugins() {
-			return plugins;
-		}
-
-		protected <T> Stream<T> stream(Class<T> type) {
-			Stream<T> stream = stream().filter(type::isInstance)
-				.map(type::cast);
-			return stream;
-		}
-
-		@Override
-		public <T> T getPlugin(Class<T> type) {
-			Optional<T> first = stream(type).findFirst();
-			return first.orElse(null);
-		}
-
-		@Override
-		public <T> List<T> getPlugins(Class<T> type) {
-			List<T> list = stream(type).collect(toList());
-			return list;
-		}
-
-		@Override
-		public boolean add(Object plugin) {
-			return plugins().add(plugin);
-		}
-
-		@Override
-		public boolean addAll(Collection<? extends Object> collection) {
-			return plugins().addAll(collection);
-		}
-
-		@Override
-		public boolean remove(Object plugin) {
-			return plugins().remove(plugin);
-		}
-
-		@Override
-		public Iterator<Object> iterator() {
-			return plugins().iterator();
-		}
-
-		@Override
-		public Spliterator<Object> spliterator() {
-			return plugins().spliterator();
-		}
-
-		@Override
-		public Stream<Object> stream() {
-			return plugins().stream();
-		}
-
-		@Override
-		public int size() {
-			return plugins().size();
-		}
-
-		@Override
-		public String toString() {
-			return plugins().toString();
-		}
-
-		/**
-		 * Magic to load the plugins. This is quite tricky actually since we
-		 * allow plugins to be downloaded (this is mainly intended for
-		 * repositories since in general plugins should use extensions, however
-		 * to bootstrap the extensions we need more). Since downloads might need
-		 * plugins for passwords and protocols we need to first load the paths
-		 * specified on the plugin clause, then check if there are any local
-		 * plugins (starting with aQute.bnd and be able to load from our own
-		 * class loader).
-		 * <p>
-		 * After that, we load the plugin paths, these can use the built in
-		 * connectors.
-		 * <p>
-		 * Last but not least, we load the remaining plugins.
-		 */
-		protected void loadPlugins(String pluginString, String pluginPathString) {
-			Parameters pluginParameters = new Parameters(pluginString, Processor.this, true);
-			CL loader = getLoader();
-
-			// First add the plugin-specific paths from their path: directives
-			for (Entry<String, Attrs> entry : pluginParameters.entrySet()) {
-				String key = removeDuplicateMarker(entry.getKey());
-				String path = entry.getValue()
-					.get(PATH_DIRECTIVE);
-				if (path != null) {
-					String parts[] = path.split("\\s*,\\s*");
-					try {
-						for (String p : parts) {
-							File f = getFile(p).getAbsoluteFile();
-							loader.add(f);
-						}
-					} catch (Exception e) {
-						error("Problem adding path %s to loader for plugin %s. Exception: (%s)", path, key, e);
-					}
-				}
-			}
-
-			//
-			// Try to load any plugins that are local
-			// these must start with aQute.bnd.* and
-			// and be possible to load. The main intention
-			// of this code is to load the URL connectors so that
-			// any access to remote plugins can use the connector
-			// model.
-			//
-
-			Set<String> loaded = new HashSet<>();
-			for (Entry<String, Attrs> entry : pluginParameters.entrySet()) {
-				String className = removeDuplicateMarker(entry.getKey());
-				Attrs attrs = entry.getValue();
-
-				logger.debug("Trying pre-plugin {}", className);
-
-				Object plugin = loadPlugin(Processor.this.getClass()
-					.getClassLoader(), attrs, className, true);
-				if (plugin != null) {
-					// with the marker!!
-					loaded.add(entry.getKey());
-					add(plugin);
-				}
-			}
-
-			//
-			// Make sure we load each plugin only once
-			// by removing the entries that were successfully loaded
-			//
-			pluginParameters.keySet()
-				.removeAll(loaded);
-
-			loadPluginPath(pluginPathString, loader);
-
-			//
-			// Load the remaining plugins
-			//
-
-			for (Entry<String, Attrs> entry : pluginParameters.entrySet()) {
-				String className = removeDuplicateMarker(entry.getKey());
-				Attrs attrs = entry.getValue();
-
-				logger.debug("Loading secondary plugin {}", className);
-
-				// We can defer the error if the plugin specifies
-				// a command name. In that case, we'll verify that
-				// a bnd file does not contain any references to a
-				// plugin
-				// command. The reason this feature was added was
-				// to compile plugin classes with the same build.
-				String commands = attrs.get(COMMAND_DIRECTIVE);
-
-				Object plugin = loadPlugin(loader, attrs, className, commands != null);
-				if (plugin != null)
-					add(plugin);
-				else {
-					if (commands == null)
-						error("Cannot load the plugin %s", className);
-					else {
-						Collection<String> cs = split(commands);
-						missingCommand.addAll(cs);
-					}
-				}
-			}
-		}
-
-		/**
-		 * Add the @link {@link Constants#PLUGINPATH} entries (which are file
-		 * names) to the class loader. If this file does not exist, and there is
-		 * a {@link Constants#PLUGINPATH_URL_ATTR} attribute then we download it
-		 * first from that url. You can then also specify a
-		 * {@link Constants#PLUGINPATH_SHA1_ATTR} attribute to verify the file.
-		 *
-		 * @see PLUGINPATH
-		 * @param pluginPath the clauses for the plugin path
-		 * @param loader The class loader to extend
-		 */
-		private void loadPluginPath(String pluginPath, CL loader) {
-			Parameters pluginPathParameters = new Parameters(pluginPath, Processor.this);
-			HttpClient client = null;
-			try {
-				nextClause: for (Entry<String, Attrs> entry : pluginPathParameters.entrySet()) {
-
-					File f = getFile(entry.getKey()).getAbsoluteFile();
-					if (!f.isFile()) {
-
-						//
-						// File does not exist! Check if we need to download
-						//
-
-						String url = entry.getValue()
-							.get(PLUGINPATH_URL_ATTR);
-						if (url != null) {
-							try {
-								logger.debug("downloading {} to {}", url, f.getAbsoluteFile());
-								URL u = new URL(url);
-								if (client == null) {
-									@SuppressWarnings("resource")
-									HttpClient c = new HttpClient();
-									c.setRegistry(this);
-									c.readSettings(Processor.this);
-
-									//
-									// Allow the URLConnectionHandlers to
-									// interact
-									// with the
-									// connection so they can sign it or
-									// decorate it
-									// with
-									// a password etc.
-									//
-									stream(URLConnectionHandler.class)
-										.forEach(c::addURLConnectionHandler);
-									client = c;
-								}
-
-								//
-								// Copy the url to the file
-								//
-								IO.mkdirs(f.getParentFile());
-								try (Resource resource = Resource.fromURL(u, client)) {
-									try (OutputStream out = IO.outputStream(f)) {
-										resource.write(out);
-									}
-									long lastModified = resource.lastModified();
-									if (lastModified > 0L) {
-										f.setLastModified(lastModified);
-									}
-								}
-
-								//
-								// If there is a sha specified, we verify the
-								// download
-								// of the
-								// the file.
-								//
-								String digest = entry.getValue()
-									.get(PLUGINPATH_SHA1_ATTR);
-								if (digest != null) {
-									if (Hex.isHex(digest.trim())) {
-										byte[] sha1 = Hex.toByteArray(digest);
-										byte[] filesha1 = SHA1.digest(f)
-											.digest();
-										if (!Arrays.equals(sha1, filesha1)) {
-											error(
-												"Plugin path: %s, specified url %s and a sha1 but the file does not match the sha",
-												entry.getKey(), url);
-										}
-									} else {
-										error(
-											"Plugin path: %s, specified url %s and a sha1 '%s' but this is not a hexadecimal",
-											entry.getKey(), url, digest);
-									}
-								}
-							} catch (Exception e) {
-								exception(e, "Failed to download plugin %s from %s, error %s", entry.getKey(), url, e);
-								continue nextClause;
-							}
-						} else {
-							error("No such file %s from %s and no 'url' attribute on the path so it can be downloaded",
-								entry.getKey(), this);
-							continue nextClause;
-						}
-					}
-					logger.debug("Adding {} to loader for plugins", f);
-					loader.add(f);
-				}
-			} finally {
-				IO.close(client);
-			}
-		}
-
-		/**
-		 * Load a plugin and customize it. If the plugin cannot be loaded then
-		 * we return null.
-		 *
-		 * @param loader Name of the loader
-		 * @param attrs
-		 * @param className
-		 */
-		private Object loadPlugin(ClassLoader loader, Attrs attrs, String className, boolean ignoreError) {
-			try {
-				Class<?> c = loader.loadClass(className);
-				Object plugin = publicLookup().findConstructor(c, defaultConstructor)
-					.invoke();
-				customize(plugin, attrs, this);
-				if (plugin instanceof AutoCloseable) {
-					closeablePlugins.add((AutoCloseable) plugin);
-				}
-				return plugin;
-			} catch (NoClassDefFoundError e) {
-				if (!ignoreError)
-					exception(e, "Failed to load plugin %s;%s, error: %s ", className, attrs, e);
-			} catch (ClassNotFoundException e) {
-				if (!ignoreError)
-					exception(e, "Failed to load plugin %s;%s, error: %s ", className, attrs, e);
-			} catch (Error e) {
-				throw e;
-			} catch (Throwable e) {
-				exception(e, "Unexpected error loading plugin %s-%s: %s", className, attrs, e);
-			}
-			return null;
-		}
-
-		boolean isMissingPlugin(String name) {
-			return missingCommand.contains(name);
-		}
-
-		protected void close() {
-			closeablePlugins.forEach(IO::close);
-			closeablePlugins.clear();
-			plugins.clear();
-			missingCommand.clear();
-		}
-	}
-
 
 	/**
 	 * Is called after the PluginsContainer is initialized.
@@ -959,8 +577,6 @@ public class Processor extends Domain implements Reporter, Registry, Constants, 
 	protected void addExtensions(PluginsContainer pluginsContainer) {
 
 	}
-
-	static final MethodType defaultConstructor = methodType(void.class);
 
 	protected void setTypeSpecificPlugins(PluginsContainer pluginsContainer) {
 		pluginsContainer.add(getExecutor());
